refactor: avoid form god component
